fix bugs and code style
