+++ tests/test_full_screen.py
@@ -1,12 +1,13 @@
 import asyncio
 
+from browser_use.browser.browser import BrowserConfig
 from browser_use.driver import Driver
 
 
 async def test_full_screen(start_fullscreen: bool, maximize: bool):
-	async with Driver("playwright") as p:
+	async with Driver('playwright', config=BrowserConfig()) as p:
 		assert p.chromium is not None
-		browser = await p.chromium.launch(
+		browser = await p.chromium.open(
 			headless=False,
 			args=['--start-maximized'],
 		)
